Incluindo classe DTO e exceptions personalizadas

+++ OneDrive/Área de Trabalho/Projetos/Java/AppContatos/src/main/java/br/com/Felipe/App/Contatos/resource/ContatoResource.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import br.com.Felipe.App.Contatos.model.Contato;
 import br.com.Felipe.App.Contatos.service.ContatoService;
+import io.swagger.v3.oas.annotations.Operation;
 
 @RestController
 @RequestMapping("/api/contatos")
@@ -28,9 +29,11 @@ public class ContatoResource {
 		this.contatoService = contatoService;
 	}
 	
+	@Operation(summary = "Cadastrar um novo contato para uma pessoa")
 	@PostMapping("/save")
 	public ResponseEntity<Contato> save(@RequestBody Contato contato){
 		Contato newContato = contatoService.save(contato);
+		
 		if(newContato == null) {
 			return ResponseEntity.notFound().build();
 		}
@@ -38,9 +41,11 @@ public class ContatoResource {
 		return ResponseEntity.ok(newContato);
 	}
 	
+	@Operation(summary = "Buscar os dados de contato de uma pessoa por ID")
 	@RequestMapping("/{id}")
 	public ResponseEntity<Optional<Contato>> getById(@PathVariable Long id){
 		Optional<Contato> findContato = contatoService.getById(id);
+		
 		if(findContato == null) {
 			return ResponseEntity.notFound().build();
 		}
@@ -52,9 +57,11 @@ public class ContatoResource {
 		return ResponseEntity.ok(findContato);
 	}
 	
+	@Operation(summary = "Buscar todos os dados de contato cadastrados")
 	@RequestMapping("/all")
 	public ResponseEntity<List<Contato>> getAll(){
 		List<Contato> findContato = contatoService.getAll();
+		
 		if(findContato == null) {
 			return ResponseEntity.notFound().build();
 		}
@@ -65,9 +72,11 @@ public class ContatoResource {
 		return ResponseEntity.ok(findContato);
 	}
 	
+	@Operation(summary = "Atualizar os dados de contato de uma pessoa")
 	@PutMapping("/update")
 	public ResponseEntity<Contato> update(@RequestBody Contato contato){
 		Contato upContato = contatoService.update(contato);
+		
 		if(upContato == null) {
 			return ResponseEntity.notFound().build();
 		}
@@ -75,9 +84,21 @@ public class ContatoResource {
 		return ResponseEntity.ok(upContato);		
 	}
 	
+	@Operation(summary = "Remover os dados de contato de uma pessoa por ID")
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<Contato> delete(@PathVariable Long id){
-		contatoService.delete(id);		
+		Optional<Contato> deletar = contatoService.getById(id);
+		
+		if(deletar == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		if(deletar.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		contatoService.delete(id);
+		
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 }
