+++ OneDrive/Área de Trabalho/Projetos/Java/AppContatos/src/main/java/br/com/Felipe/App/Contatos/service/ContatoService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.Felipe.App.Contatos.exception.ResourceNotFoundException;
 import br.com.Felipe.App.Contatos.interfaces.ContatoServiceInterface;
 import br.com.Felipe.App.Contatos.model.Contato;
 import br.com.Felipe.App.Contatos.model.Pessoa;
@@ -26,36 +27,82 @@ public class ContatoService implements ContatoServiceInterface {
 	}
 	
 	@Override
-	public Contato save(Contato contato) {
+	public Contato save(Contato contato) throws ResourceNotFoundException  {
 		
 		if(contato.getPessoa().getId() != null) {
 			Optional<Pessoa> findPessoa = pessoaRepository.findById(contato.getPessoa().getId());
 			if(findPessoa.isEmpty()) {
-				System.out.println("Pessoa não encontrado.");
-				return null;
+				throw new ResourceNotFoundException("[CONTATO] Pessoa não encontrada");
 			}else {
+				if(contato.getTipoContato().toString().length() == 0) {
+					throw new ResourceNotFoundException("[CONTATO] O tipo do Contato, não pode ser vazio");
+				}
+				
+				if(contato.getContato().toString().length() == 0) {
+					throw new ResourceNotFoundException("[CONTATO] O Contato, não pode ser vazio");
+				}
 				contato.setPessoa(findPessoa.get());
 				return contatoRepository.save(contato);
 			}
 		}else {
-			System.out.println("Não existe Pessoa");
-			return null;
+			throw new ResourceNotFoundException("[CONTATO] Pessoa não encontrada");
 		}		
 	}
 
 	@Override
-	public Optional<Contato> getById(Long id) {
-		return contatoRepository.findById(id);
+	public Optional<Contato> getById(Long id) throws ResourceNotFoundException {
+		Optional<Contato> contato = contatoRepository.findById(id);
+		
+		if(contato == null) {
+			throw new ResourceNotFoundException("[CONTATO] Contato não encontrado pelo ID: " + id);
+		}
+		if(contato.isEmpty()) {
+			throw new ResourceNotFoundException("[CONTATO] Contato não encontrado pelo ID: " + id);
+		}
+		
+		return contato;
 	}
 
 	@Override
-	public List<Contato> getAll() {
-		return contatoRepository.findAll();
+	public List<Contato> getAll() throws ResourceNotFoundException {
+		List<Contato> listContato = contatoRepository.findAll();
+		
+		if(listContato == null) {
+			throw new ResourceNotFoundException("[CONTATO] Nenhuma contato cadastrado");
+			}
+			
+		if(listContato.size() == 0) {
+			throw new ResourceNotFoundException("[CONTATO] Nenhuma contato cadastrado");
+			}
+			
+		return listContato;
 	}
 
 	@Override
-	public Contato update(Contato contato) {
+	public Contato update(Contato contato) throws ResourceNotFoundException {
 		Optional<Contato> findContato = contatoRepository.findById(contato.getId());
+		Optional<Pessoa> findPessoa = pessoaRepository.findById(contato.getPessoa().getId());
+		
+		if(findContato == null) {
+			throw new ResourceNotFoundException("[CONTATO] Contato não encontrado pelo ID: " + contato.getId());
+		}
+		
+		if(findContato.isEmpty()) {
+			throw new ResourceNotFoundException("[CONTATO] Contato não encontrado pelo ID: " + contato.getId());
+		}
+		
+		if(findPessoa.isEmpty()) {
+			throw new ResourceNotFoundException("[CONTATO] Pessoa não encontrada pelo ID: " + contato.getPessoa().getId());
+		}
+		
+		if(contato.getTipoContato().toString().length() == 0) {
+			throw new ResourceNotFoundException("[CONTATO] O tipo do Contato, não pode ser vazio");
+		}
+		
+		if(contato.getContato().toString().length() == 0) {
+			throw new ResourceNotFoundException("[CONTATO] O Contato, não pode ser vazio");
+		}
+		
 		if(findContato.isPresent()) {
 			Contato upContato = findContato.get();
 			upContato.setPessoa(findContato.get().getPessoa());
@@ -67,7 +114,16 @@ public class ContatoService implements ContatoServiceInterface {
 	}
 
 	@Override
-	public void delete(Long id) {
+	public void delete(Long id) throws ResourceNotFoundException {
+		Optional<Contato> deletar = contatoRepository.findById(id);
+		
+		if(deletar == null) {
+			throw new ResourceNotFoundException("[CONTATO] ID: " + id + " não encontrado, não é possível deletar");
+		}
+		if(deletar.isEmpty()) {
+			throw new ResourceNotFoundException("[CONTATO] ID: " + id + " não encontrado, não é possível deletar");
+		}
+		
 		contatoRepository.deletePorId(id);
 	}
